refactor: 위치 도메인 패키지 변경
