+++ src/main/java/controller/JanggiGame.java
@@ -1,0 +1,4 @@
+package controller;
+
+public class JanggiGame {
+}
