+++ src/main/java/domain/piece/Blank.java
@@ -1,6 +1,7 @@
 package domain.piece;
 
 import domain.Team;
+import domain.position.Position;
 
 public class Blank extends Piece {
 
@@ -8,4 +9,8 @@ public class Blank extends Piece {
         super(Team.NONE);
     }
 
+    @Override
+    public boolean canMove(Position currentPosition, Position targetPosition, PieceProvider pieceProvider) {
+        return false;
+    }
 }
