+++ src/main/java/domain/piece/Car.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.position.Position;
 import domain.Team;
 import domain.strategy.CarStrategy;
 import domain.strategy.MoveStrategy;
@@ -15,9 +15,9 @@ public class Car extends Piece {
         this.moveStrategy = new CarStrategy();
     }
 
+    @Override
     public boolean canMove(Position currentPosition, Position targetPosition, PieceProvider pieceProvider) {
         List<Position> moveCandidates = moveStrategy.getMoveCandidates(currentPosition, pieceProvider);
-
         for (Position candidatePosition : moveCandidates) {
             if (candidatePosition.equals(targetPosition)) {
                 return true;
