+++ src/main/java/domain/piece/Elephant.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.position.Position;
 import domain.Team;
 import domain.strategy.ElephantStrategy;
 import domain.strategy.MoveStrategy;
@@ -15,9 +15,9 @@ public class Elephant extends Piece {
         this.moveStrategy = new ElephantStrategy();
     }
 
+    @Override
     public boolean canMove(Position currentPosition, Position targetPosition, PieceProvider pieceProvider) {
         List<Position> moveCandidates = moveStrategy.getMoveCandidates(currentPosition, pieceProvider);
-
         for (Position candidatePosition : moveCandidates) {
             if (candidatePosition.equals(targetPosition)) {
                 return true;
