+++ src/main/java/domain/piece/Guard.java
@@ -1,6 +1,6 @@
 package domain.piece;
 
-import domain.Position;
+import domain.position.Position;
 import domain.Team;
 import domain.strategy.MoveStrategy;
 import domain.strategy.PalaceStrategy;
@@ -15,10 +15,10 @@ public class Guard extends Piece {
         this.moveStrategy = new PalaceStrategy();
     }
 
+    @Override
     public boolean canMove(Position currentPosition, Position targetPosition, PieceProvider pieceProvider) {
         List<Position> moveCandidates = moveStrategy.getMoveCandidates(currentPosition, pieceProvider);
         boolean isTargetPositionBlank = pieceProvider.isBlank(targetPosition);
-
         for (Position candidatePosition : moveCandidates) {
             if (candidatePosition.equals(targetPosition) && isTargetPositionBlank) {
                 return true;
