+++ src/main/java/domain/strategy/PawnStrategy.java
@@ -1,6 +1,6 @@
 package domain.strategy;
 
-import domain.Position;
+import domain.position.Position;
 import domain.piece.PieceProvider;
 
 import java.util.ArrayList;
