+++ src/test/java/janggiBoard/BoardInitialTest.java
@@ -1,7 +1,7 @@
 package janggiBoard;
 
 import domain.JanggiBoard;
-import domain.Position;
+import domain.position.Position;
 import domain.Team;
 import domain.piece.*;
 import org.junit.jupiter.api.BeforeEach;
