+++ src/test/java/janggiBoard/strategyTest/HorseStrategyTest.java
@@ -1,7 +1,7 @@
 package janggiBoard.strategyTest;
 
 
-import domain.Position;
+import domain.position.Position;
 import domain.piece.Blank;
 import domain.piece.Piece;
 import domain.piece.PieceProvider;
